requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==2.0.3
djangorestframework==3.9.1
idna==2.8
mysqlclient==1.3.12
pytz==2018.3
requests==2.21.0
urllib3==1.24.1
